test: DCI 패턴을 적용한 CouponEventTest 적용
