+++ backend/src/test/java/com/woowacourse/kkogkkog/domain/CouponEventTest.java
@@ -5,117 +5,145 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import com.woowacourse.kkogkkog.exception.ForbiddenException;
 import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
+@DisplayName("CouponEvent 클래스의")
 class CouponEventTest {
 
-    @Test
-    @DisplayName("쿠폰을 받은 사람이 쿠폰 사용 요청을 보낼 수 있다.")
-    void receiverCanRequest() {
-        boolean isSender = false;
-        boolean isReceiver = true;
+    @Nested
+    @DisplayName("REQUEST 이벤트는")
+    class Request {
 
-        assertThatNoException()
-            .isThrownBy(() -> CouponEvent.REQUEST.checkExecutable(isSender, isReceiver));
+        @Test
+        @DisplayName("받은 사람이 보낼 수 있다.")
+        void success() {
+            boolean isSender = false;
+            boolean isReceiver = true;
+
+            assertThatNoException()
+                .isThrownBy(() -> CouponEvent.REQUEST.checkExecutable(isSender, isReceiver));
+        }
+
+        @Test
+        @DisplayName("보낸 사람이 보내면, 예외를 던진다.")
+        void fail_senderRequest() {
+            boolean isSender = true;
+            boolean isReceiver = false;
+
+            assertThatThrownBy(() -> CouponEvent.REQUEST.checkExecutable(isSender, isReceiver))
+                .isInstanceOf(ForbiddenException.class);
+        }
     }
 
-    @Test
-    @DisplayName("쿠폰을 보낸 사람이 쿠폰 사용 요청을 보내는 경우 예외가 발생한다.")
-    void senderCanNotRequest() {
-        boolean isSender = true;
-        boolean isReceiver = false;
+    @Nested
+    @DisplayName("CANCEL 이벤트는")
+    class Cancel {
 
-        assertThatThrownBy(() -> CouponEvent.REQUEST.checkExecutable(isSender, isReceiver))
-            .isInstanceOf(ForbiddenException.class);
+        @Test
+        @DisplayName("받은 사람이 보낼 수 있다.")
+        void success() {
+            boolean isSender = false;
+            boolean isReceiver = true;
+
+            assertThatNoException()
+                .isThrownBy(() -> CouponEvent.CANCEL.checkExecutable(isSender, isReceiver));
+        }
+
+        @Test
+        @DisplayName("보낸 사람이 보내면, 예외를 던진다.")
+        void fail_senderCancel() {
+            boolean isSender = true;
+            boolean isReceiver = false;
+
+            assertThatThrownBy(() -> CouponEvent.CANCEL.checkExecutable(isSender, isReceiver))
+                .isInstanceOf(ForbiddenException.class);
+        }
     }
 
-    @Test
-    @DisplayName("쿠폰을 받은 사람이 쿠폰 사용 요청을 취소할 수 있다.")
-    void receiverCanCancel() {
-        boolean isSender = false;
-        boolean isReceiver = true;
+    @Nested
+    @DisplayName("DECLINE 이벤트는")
+    class Decline {
 
-        assertThatNoException()
-            .isThrownBy(() -> CouponEvent.CANCEL.checkExecutable(isSender, isReceiver));
+        @Test
+        @DisplayName("보낸 사람이 보낼 수 있다.")
+        void success() {
+            boolean isSender = true;
+            boolean isReceiver = false;
+
+            assertThatNoException()
+                .isThrownBy(() -> CouponEvent.DECLINE.checkExecutable(isSender, isReceiver));
+        }
+
+        @Test
+        @DisplayName("받은 사람이 보내면, 예외를 던진다.")
+        void fail_receiverDecline() {
+            boolean isSender = false;
+            boolean isReceiver = true;
+
+            assertThatThrownBy(() -> CouponEvent.DECLINE.checkExecutable(isSender, isReceiver))
+                .isInstanceOf(ForbiddenException.class);
+        }
     }
 
-    @Test
-    @DisplayName("쿠폰을 보낸 사람이 쿠폰 사용 요청 취소를 보내는 경우 예외가 발생한다.")
-    void senderCanNotCancel() {
-        boolean isSender = true;
-        boolean isReceiver = false;
 
-        assertThatThrownBy(() -> CouponEvent.CANCEL.checkExecutable(isSender, isReceiver))
-            .isInstanceOf(ForbiddenException.class);
+    @Nested
+    @DisplayName("ACCEPT 이벤트는")
+    class Accept {
+
+        @Test
+        @DisplayName("보낸 사람이 보낼 수 있다.")
+        void senderCanAccept() {
+            boolean isSender = true;
+            boolean isReceiver = false;
+
+            assertThatNoException()
+                .isThrownBy(() -> CouponEvent.ACCEPT.checkExecutable(isSender, isReceiver));
+        }
+
+        @Test
+        @DisplayName("받은 사람이 보내면, 예외를 던진다.")
+        void receiverCanNotAccept() {
+            boolean isSender = false;
+            boolean isReceiver = true;
+
+            assertThatThrownBy(() -> CouponEvent.ACCEPT.checkExecutable(isSender, isReceiver))
+                .isInstanceOf(ForbiddenException.class);
+        }
     }
 
-    @Test
-    @DisplayName("쿠폰을 보낸 사람이 사용 요청을 거절할 수 있다.")
-    void senderCanDecline() {
-        boolean isSender = true;
-        boolean isReceiver = false;
+    @Nested
+    @DisplayName("FINISH 이벤트는")
+    class Finish {
 
-        assertThatNoException()
-            .isThrownBy(() -> CouponEvent.DECLINE.checkExecutable(isSender, isReceiver));
-    }
+        @Test
+        @DisplayName("보낸 사람이 보낼 수 있다.")
+        void senderCanFinish() {
+            boolean isSender = true;
+            boolean isReceiver = false;
 
-    @Test
-    @DisplayName("쿠폰을 받은 사람이 쿠폰 사용 요청 거절을 보내는 경우 예외가 발생한다.")
-    void receiverCanNotDecline() {
-        boolean isSender = false;
-        boolean isReceiver = true;
+            assertThatNoException()
+                .isThrownBy(() -> CouponEvent.FINISH.checkExecutable(isSender, isReceiver));
+        }
 
-        assertThatThrownBy(() -> CouponEvent.DECLINE.checkExecutable(isSender, isReceiver))
-            .isInstanceOf(ForbiddenException.class);
-    }
+        @Test
+        @DisplayName("받은 사람이 보낼 수 있다.")
+        void receiverCanFinish() {
+            boolean isSender = false;
+            boolean isReceiver = true;
 
-    @Test
-    @DisplayName("쿠폰을 보낸 사람이 사용 요청을 수락할 수 있다.")
-    void senderCanAccept() {
-        boolean isSender = true;
-        boolean isReceiver = false;
+            assertThatNoException()
+                .isThrownBy(() -> CouponEvent.FINISH.checkExecutable(isSender, isReceiver));
+        }
 
-        assertThatNoException()
-            .isThrownBy(() -> CouponEvent.ACCEPT.checkExecutable(isSender, isReceiver));
-    }
+        @Test
+        @DisplayName("쿠폰과 관계없는 사람이 보내면, 예외를 던진다.")
+        void otherCanNotFinish() {
+            boolean isSender = false;
+            boolean isReceiver = false;
 
-    @Test
-    @DisplayName("쿠폰을 받은 사람이 쿠폰 사용 요청 수락을 보내는 경우 예외가 발생한다.")
-    void receiverCanNotAccept() {
-        boolean isSender = false;
-        boolean isReceiver = true;
-
-        assertThatThrownBy(() -> CouponEvent.ACCEPT.checkExecutable(isSender, isReceiver))
-            .isInstanceOf(ForbiddenException.class);
-    }
-
-    @Test
-    @DisplayName("쿠폰을 보낸 사람이 사용 완료를 할 수 있다")
-    void senderCanFinish() {
-        boolean isSender = true;
-        boolean isReceiver = false;
-
-        assertThatNoException()
-            .isThrownBy(() -> CouponEvent.FINISH.checkExecutable(isSender, isReceiver));
-    }
-
-    @Test
-    @DisplayName("쿠폰을 받은 사람이 사용 완료를 할 수 있다")
-    void receiverCanFinish() {
-        boolean isSender = false;
-        boolean isReceiver = true;
-
-        assertThatNoException()
-            .isThrownBy(() -> CouponEvent.FINISH.checkExecutable(isSender, isReceiver));
-    }
-
-    @Test
-    @DisplayName("쿠폰을 보내거나 받지 않은 사람이 쿠폰 사용 완료를 보내는 경우 예외가 발생한다.")
-    void otherCanNotFinish() {
-        boolean isSender = false;
-        boolean isReceiver = false;
-
-        assertThatThrownBy(() -> CouponEvent.FINISH.checkExecutable(isSender, isReceiver))
-            .isInstanceOf(ForbiddenException.class);
+            assertThatThrownBy(() -> CouponEvent.FINISH.checkExecutable(isSender, isReceiver))
+                .isInstanceOf(ForbiddenException.class);
+        }
     }
 }
